// src/app/models/pdpa-privacy/pdpa-privacy.ts
export interface PdpaPrivacyModel {
    id: number;
    ppCode: string;
    status: number;
    version: string;
    titleTh: string;
    titleEn: string;
    titleZh: string;
    descriptionTh: string;
    descriptionEn: string;
    descriptionZh: string;
    creatorUserId: number;
    creationTime: string;
    lastModifierUserId: number;
    lastModificationTime: string;
    isDeleted: number;
}